maskfit/requirements.txt
accelerate==0.30.1
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
datasets==2.19.1
diffusers==0.27.2
dill==0.3.8
docker-pycreds==0.4.0
filelock==3.14.0
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
huggingface-hub==0.23.0
idna==3.7
importlib_metadata==7.1.0
Jinja2==3.1.4
kiwisolver==1.4.5
lightning==2.3.0
lightning-utilities==0.11.2
MarkupSafe==2.1.5
matplotlib==3.9.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
packaging==24.0
pandas==2.2.2
pillow==10.3.0
platformdirs==4.2.2
protobuf==5.27.1
psutil==5.9.8
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytorch-lightning==2.2.4
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.1
safetensors==0.4.3
sentry-sdk==2.5.1
setproctitle==1.3.3
setuptools==69.5.1
six==1.16.0
smmap==5.0.1
sympy==1.12
timm==1.0.3
torch==2.3.0
torchmetrics==1.4.0.post0
torchvision==0.18.0
tqdm==4.66.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wandb==0.17.2
wheel==0.43.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.18.2